+++ crates/jlrs/src/lib.rs
@@ -196,8 +196,13 @@
 //!
 //!   Link with a debug build of Julia on Linux.
 //!
-//! You can enable all features except `debug`, `i686`, `windows`, and `lto` by enabling the
-//! `full` feature. If you don't want to enable any runtimes either, you can use `full-no-rt`.
+//! - `static-cache`
+//!
+//!   Experimental, use static-generics crate to cache constructed types.
+//!
+//! You can enable all features except `debug`, `i686`, `windows`, `static-cache`, and `lto` by
+//! enabling the `full` feature. If you don't want to enable any runtimes either, you can use
+//! `full-no-rt`.
 //!
 //!
 //! ## Environment variables